% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{data_0_a}
\alias{data_0_a}
\alias{data_0_b}
\alias{data_1}
\alias{data_2}
\alias{data_3}
\alias{data_4}
\alias{data_5}
\alias{data_6}
\alias{data_7}
\alias{data_8}
\alias{data_9}
\alias{data_10}
\alias{data_11}
\alias{data_44}
\alias{data_100}
\alias{data_200}
\alias{data_300}
\title{Simulated BAAR benchmarking datasets}
\format{
A data frame with two columns:
\describe{
\item{time}{Integer sequence denoting the observation index.}
\item{\if{html}{\out{<dataset>}}}{Numeric observations for the simulated series. The column
name matches the dataset object (e.g., \code{data_10}).}
}

An object of class \code{data.frame} with 90 rows and 2 columns.

An object of class \code{data.frame} with 90 rows and 2 columns.

An object of class \code{data.frame} with 90 rows and 2 columns.

An object of class \code{data.frame} with 90 rows and 2 columns.

An object of class \code{data.frame} with 90 rows and 2 columns.

An object of class \code{data.frame} with 90 rows and 2 columns.

An object of class \code{data.frame} with 90 rows and 2 columns.

An object of class \code{data.frame} with 90 rows and 2 columns.

An object of class \code{data.frame} with 90 rows and 2 columns.

An object of class \code{data.frame} with 90 rows and 2 columns.

An object of class \code{data.frame} with 90 rows and 2 columns.

An object of class \code{data.frame} with 90 rows and 2 columns.

An object of class \code{data.frame} with 300 rows and 2 columns.

An object of class \code{data.frame} with 900 rows and 2 columns.

An object of class \code{data.frame} with 900 rows and 2 columns.

An object of class \code{data.frame} with 200 rows and 2 columns.
}
\source{
Generated via functions in \code{\link{test_data_generators}}.
}
\usage{
data(data_0_a)

data(data_0_b)

data(data_1)

data(data_2)

data(data_3)

data(data_4)

data(data_5)

data(data_6)

data(data_7)

data(data_8)

data(data_9)

data(data_10)

data(data_11)

data(data_44)

data(data_100)

data(data_200)

data(data_300)

data_0_b

data_1

data_2

data_3

data_4

data_5

data_6

data_7

data_8

data_9

data_10

data_11

data_44

data_100

data_200

data_300
}
\description{
Collection of synthetic time-series data frames used to validate and
benchmark the Bayesian Adaptive Auto-Regression (BAAR) sampler. Each dataset
generated from the helper functions documented in
\code{\link{test_data_generators}} and stored for reuse in
examples/vignettes. Every object contains a \code{time} column of sequential
indices and a second column named after the dataset (for example
\code{data_0_a}).
}
\examples{
data(data_0_a)
head(data_0_a)
}
\keyword{datasets}
