% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\name{test_data_generators}
\alias{test_data_generators}
\alias{test_data_0_a}
\alias{test_data_0_b}
\alias{test_data_1}
\alias{test_data_2}
\alias{test_data_3}
\alias{test_data_4}
\alias{test_data_5}
\alias{test_data_6}
\alias{test_data_7}
\alias{test_data_8}
\alias{test_data_9}
\alias{test_data_10}
\alias{test_data_11}
\alias{test_data_44}
\alias{test_data_100}
\alias{test_data_200}
\alias{test_data_300}
\title{Simulated datasets used in BAAR development}
\usage{
test_data_0_a()

test_data_0_b()

test_data_1()

test_data_2()

test_data_3()

test_data_4()

test_data_5()

test_data_6()

test_data_7()

test_data_8()

test_data_9()

test_data_10()

test_data_11()

test_data_44()

test_data_100()

test_data_200()

test_data_300()
}
\value{
A data frame with simulated time-series observations.
}
\description{
These helper functions generate synthetic time-series examples that were
used during the development of the Bayesian Adaptive Auto-Regression (BAAR)
sampler. Each function returns a data frame with a \code{time} column and a
second column containing the simulated response values.
}
